requirements.txt
azure-identity
azure-storage-blob
python-dotenv
requests
azure-ai-contentunderstanding